+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog], and this project adheres to [Semantic Versioning].
 
+## [Unreleased]
+
+### Fixed
+
+* codegen: Untyped originations are now correctly handled.
+* index: Do not deserialize originations matched by the `source` filter.
+
 ## [6.3.0] - 2022-11-15
 
 ### Added